add search func to our app
we are passing state from app to both List and Search Component.
actually this was pretty neat, and with practice this approach will
become a second nature to me i am sure.

this commits are for personal use. so don't mad if you don't understand

if you have question just `git log` and send me email